+++ CHANGELOG.md
@@ -27,3 +27,7 @@
 - added support for application/hal+json MIME-type
 - added ModelBuilder for comma serperated arrays
 - changed exception logging to log exception ToString() by default, exception object logged optionally using settings
+
+## 3.0.3
+
+- added ability to disable the global exception filter